// shared/src/types/tmdb.ts
export type TMDBVideo = {
  id: string
  key: string
  name: string
  site: string
  type: string
  official: boolean
  iso_639_1: string
}

export type TMDBVideosResult = {
  results: TMDBVideo[]
}